+++ common/src/main/java/mtr/sound/bve/MotorDataBase.java
@@ -4,8 +4,8 @@ public abstract class MotorDataBase {
 
     public abstract int getSoundCount();
 
-    public abstract float getPitch(int index, float speed, int accel);
+    public abstract float getPitch(int index, float speed, float accel);
 
-    public abstract float getVolume(int index, float speed, int accel);
+    public abstract float getVolume(int index, float speed, float accel);
 
 }
